requirements.txt
# DeepGuard Multi-Agentic Deepfake Detection System
# Dependencies

# Core Framework
langgraph==0.2.34
langchain==0.3.1
langchain-anthropic==0.2.1
langchain-community==0.3.0

# Computer Vision & Deep Learning
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
pillow==10.1.0
facenet-pytorch==2.5.3
timm==0.9.12

# Explainability
grad-cam==1.4.8
shap==0.43.0
captum==0.7.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3

# API & Web
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
